Fixes remaining references to key codes, starts some text-proc reorg.
