+++ web/package.json
@@ -22,8 +22,8 @@
     "google-closure-compiler": "^20171203.0.0",
     "google-closure-library": "^20171203.0.0",
     "html2canvas": "^1.0.0-alpha.12",
-    "karma": "^3.1.1",
-    "karma-browserstack-launcher": "^1.3.0",
+    "karma": "^3.1.4",
+    "karma-browserstack-launcher": "^1.4.0",
     "karma-chai": "^0.1.0",
     "karma-chrome-launcher": "^2.2.0",
     "karma-edge-launcher": "^0.4.2",
@@ -38,7 +38,7 @@
     "karma-teamcity-reporter": "^1.1.0",
     "mocha": "^5.2.0",
     "modernizr": "^3.6.0",
-    "typescript": "^3.1.6"
+    "typescript": "^3.2.2"
   },
   "scripts": {
     "tsc": "tsc",
